pyproject.toml
[project]
name = "sflabs-ai-assistant"
version = "1.24.0"
description = "SFLabs AI assistant with agents for DevOps and SRE tasks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit==2.8.1",
    "chromadb==1.0.13",
    "fpdf==1.7.2",
    "GitPython==3.1.44",
    "langchain==0.3.27",
    "langchain-aws==0.2.25",
    "langchain-chroma==0.2.4",
    "langchain-community==0.3.27",
    "langchain-core==0.3.81",
    "langchain-google-genai==2.1.8",
    "langchain-mcp-adapters==0.1.9",
    "langgraph==0.6.2",
    "markdown2==2.5.3",
    "mermaid-py==0.8.0",
    "mcp==1.12.3",
    "mcp-server-fetch==2025.4.7",
    "pypdf>=6.0.0",
    "PyPDF2==3.0.1",
    "python-docx==1.2.0",
    "python-dotenv==1.1.0",
    "PyYAML==6.0.2",
    "reportlab==4.4.2",
    "slack-bolt==1.23.0",
    "uv==0.7.19",
    "weasyprint==65.1",
    "pandas==2.3.2",
    "matplotlib==3.10.6",
    "google-cloud-speech>=2.33.0",
    "requests==2.32.4",
    "beautifulsoup4==4.13.4",
    "botbuilder-core>=4.17.1",
    "botbuilder-schema>=4.17.1",
    "ruff>=0.12.0",
]
[dependency-groups]
code_check = [
    "bandit==1.8.6",
    "ruff==0.12.0"
]
dev = [
    "black==25.1.0",
]
